Modified the build procedure to build all libraries as shared.
The problems were caused by a misunderstanding of how make works.
If a makefile contains the rule:

a b:
	doit

it is 100% equivalent to:

a:
	doit
b:
	doit

Which means that "doit" can be executed multiple times, even if "doit"
generates the same two files.  In this case, I had rules effectively:

%.lo %.o: %.c
	./libcompile $<

However, while the libraries depended on the .lo files, the crc-gentab
program depended on the .o files, to produce a static executable.  What
would happen is that make would occasionally try to use the .o file
immediately before running libcompile again, which removes it.  Bleh.

So, this change builds everything to libtool .lo/.la files, and links
from there, but installs the private .a from the .libs subdirectory.
It's a somewhat internal feature of libtool, but it works.