Revert installed libbg-cli back to static linkage.
The CLI lib calls cli_show_help when it needs to show help text. The
library includes an implementation that dynamically generates the help
text from the options structure, but cli-generate also builds its own
function that has pre-generated text. In the dynamic .so file, this
doesn't work, since it's all pre-linked internally.

I could change cli_show_help to a weak symbol, but that adds a gcc
dependancy, and the CLI library is small and doesn't change much, so
just revert to static linkage.